fix Home Invasion cutscene bug AGAIN
bug where Mario has a chance of falling through the ground when entering the house in Home Invasion for the first time.
it wasn't fixed at all. now it seems to be.
i applied a hacky fix by checking the changes to the screen fade status and loading new collisions, but the new collisions will only be loaded if the player is indoors.