fix bug with mission 'Catalyst'
after getting on the train as Mario, you could see that the camera was moving along the train just fine, but the player character in the radar was not. this might also fix the chase scene in Reuniting the Families, and the helicopter scene in The Da Nang Thang, although this isn't confirmed.