add animations for cutscene in mission "OG Loc"
it's the cutscene when you arrive to Freddy's crib.
also redesigned the cjAnimKeys map to use a struct instead of just a function callback. this struct contains the callback AND a bool to stop the animation from looping, so that CJ's anim task doesn't set ACT_CUSTOM_ANIM_TO_ACTION again if that animation was already played once.