add vending machine animation
i can't for the life of me find a way to intercept the can that CJ is holding so it appears on Mario's right hand.
calling task->m_pEntityToHold->Setposn() does nothing.
i tried analyzing gta_sa.exe on ghidra to find function calls to 0x6940A0 (CTaskSimpleHoldEntity::SetPedPosition) and then add those calls as a custom event on main.cpp, but no dice.
so i decided to go the easy way out and just make the can invisible