weapon animations for Mario: part 1/2
Mario now has animations for guns that can be can dual-wielded: pistol, sawed-off shotgun and uzi/tec9.
part 2 will include the rest of the weapons: deagle, shotguns, mp5, rifles, etc.
melee weapons such as the baseball bat or knife don't need custom animations - they're fine as-is.

added a custom event on main.cpp that intercepts the function calls for rendering weapons on CJ's hands, so that they're rendered on Mario's hands instead.
this is the first time i use a reverse-engineering tool like ghidra to aid myself on the development of a mod, and i think i like how this is turning out