hijack CJ's GTA shadow to Mario's position
another feature made with the help of ghidra and gta-reversed-modern!
this happens after "CShadows::StoreShadowForPedObject" is called within the CPed function "CPed::PreRenderAfterTest",
which then replaces CJ's shadow entry with one placed in Mario's position.
at the moment the shadow doesn't display indoors but i'll get to that sometime later